refactor out get_zenml_dir
